+++ trainOperator/src/main/java/org/group21/trainoperator/repository/TrainRepository.java
@@ -3,5 +3,5 @@ package org.group21.trainoperator.repository;
 import org.group21.trainoperator.model.Train;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainRepository extends JpaRepository<Train, Integer> {
+public interface TrainRepository extends JpaRepository<Train, Long> {
 }
